Die when there is no such dir with translations/help.
Package-Manager: portage-2.2.0_alpha53/cvs/Linux x86_64

+++ app-office/libreoffice-l10n/libreoffice-l10n-3.4.3-r1.ebuild
@@ -1,6 +1,6 @@
 # Copyright 1999-2011 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo-x86/app-office/libreoffice-l10n/libreoffice-l10n-3.4.3-r1.ebuild,v 1.2 2011/09/06 12:51:49 scarabeus Exp $
+# $Header: /var/cvsroot/gentoo-x86/app-office/libreoffice-l10n/libreoffice-l10n-3.4.3-r1.ebuild,v 1.3 2011/09/06 19:24:43 scarabeus Exp $
 
 EAPI=4
 
@@ -75,6 +75,7 @@ src_unpack() {
 		# for english we provide just helppack, as translation is always there
 		if [[ ${lang} != en ]]; then
 			rpmdir="LibO_${PV}${RC_VERSION}_Linux_x86_langpack-rpm_${dir}/RPMS/"
+			[[ -d ${rpmdir} ]] || die "Missing directory: \"${rpmdir}\""
 			# First remove dictionaries, we want to use system ones.
 			rm -rf "${S}/${rpmdir}/"*dict*.rpm
 			rpm_unpack "./${rpmdir}/"*.rpm
@@ -82,6 +83,7 @@ src_unpack() {
 		if [[ "${LANGUAGES_HELP}" =~ "${lang}" ]] && use offlinehelp; then
 			[[ ${lang} == en ]] && dir="en-US"
 			rpmdir="LibO_${PV}${RC_VERSION}_Linux_x86_helppack-rpm_${dir}/RPMS/"
+			[[ -d ${rpmdir} ]] || die "Missing directory: \"${rpmdir}\""
 			rpm_unpack ./"${rpmdir}/"*.rpm
 		fi
 	done
